Main component renamed, index.js exports all public modules.
